pyt-image-cls-web/requirements.txt
python-magic==0.4.15 #新增
requests==2.24.0
flask==1.1.1
flask-uploads==0.2.1 #新增
Werkzeug==0.16.0     #新增
numpy==1.17.2
Pillow==6.2.0
matplotlib==3.1.1
#opencv-contrib-python==3.4.2.17
torch==1.4
torchvision==0.5.0
